inserted distribution info

+++ ubuntu.sh
@@ -26,6 +26,9 @@ do
     mkdir targets/$target || exit 1
     cp -r "huggle-$v" "targets/$target/huggle-$v" || exit 1
     cp "huggle_$v.orig.tar.gz" "targets/$target/huggle_$v.orig.tar.gz" || exit 1
+    if [ -f "targets.info/"$target"_control" ];then
+        cp "targets.info/"$target"_control" "targets/$target/huggle_$v/debian/control" || exit 1
+    fi
     cd "targets/$target/huggle-$v" || exit 1
     cat debian/changelog | sed "s/precise/$target/" | sed "s/-ppa/-$target/" > debian/.changelog || exit 1
     mv debian/.changelog debian/changelog || exit 1
